# internal/api/api.swagger.yaml
openapi: 3.0.3
info:
  title: Online Courses Platform API
  description: API для платформы онлайн-курсов (аналог Docusaurus + LMS функционал)
  version: 0.1.0
  contact:
    name: Artem
    email: artem@example.com
servers:
  - url: https://api.courses-platform.com/v1
    description: Production server
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          nullable: false
        data:
          type: object
          nullable: true
        success:
          type: boolean
          nullable: false
        requestID:
          type: string
          nullable: false
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        fullName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, instructor, admin]
        createdAt:
          type: string
          format: date-time
    UserCreate:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
...
    Section:
      type: object
      properties:
        id:
          type: integer
          format: int64
        courseId:
          type: integer
          format: int64
        title:
          type: string
        order:
          type: integer
        isFreePreview:
          type: boolean
        estimatedTime:
          type: integer
          description: minutes
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SectionCreate:
      type: object
      required: [title, order]
      properties:
        title:
          type: string
        order:
          type: integer
        isFreePreview:
          type: boolean
          default: false
        estimatedTime:
          type: integer
          description: minutes
    SectionUpdate:
      type: object
      properties:
        title:
          type: string
        order:
          type: integer
        isFreePreview:
          type: boolean
        estimatedTime:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
paths:
  /auth/register:
    post:
      operationId: authRegisterUser
      summary: Регистрация нового пользователя
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: authLoginUser
      summary: Авторизация пользователя
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
...
  /users/{userId}:
    delete:
      operationId: deleteUserById
      summary: Удалить пользователя (только для админов)
      tags: [Users, Admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Пользователь удалён
        "403":
          description: Недостаточно прав (только admin)
        "404":
          description: Пользователь не найден
...
  /courses/{courseID}/sections:
    get:
      operationId: getSections
      summary: Получить все разделы курса
      tags: [Sections]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Список разделов курса
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
        "404":
          description: Курс не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createSection
      summary: Создать новый раздел в курсе (для преподавателей/админов)
      tags: [Sections]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionCreate"
      responses:
        "201":
          description: Раздел успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
        "400":
          description: Некорректные данные (например, order уже занят)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав на создание раздела в этом курсе
        "404":
          description: Курс не найден
  /courses/{courseID}/sections/{sectionID}:
    get:
      operationId: getSectionByID
      summary: Получить информацию об одном разделе курса
      tags: [Sections]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: sectionID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Детали раздела
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
        "403":
          description: Нет доступа (раздел скрыт и пользователь не автор)
        "404":
          description: Раздел или курс не найден

    patch:
      operationId: updateSection
      summary: Обновить раздел (для преподавателей/админов)
      tags: [Sections]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: sectionID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionUpdate"
      responses:
        "200":
          description: Раздел обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
        "400":
          description: Некорректные данные
        "403":
          description: Нет прав на редактирование
        "404":
          description: Раздел или курс не найден

    delete:
      operationId: deleteSection
      summary: Удалить раздел (для преподавателей/админов)
      tags: [Sections]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: sectionID
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Раздел успешно удалён
        "403":
          description: Нет прав на удаление (например, в разделе есть уроки)
        "404":
          description: Раздел или курс не найден
...